pyproject.toml
[project]
name = "friends-birthday-reminder"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "google-api-python-client==2.190.0",
    "google-auth-httplib2==0.3.0",
    "google-auth-oauthlib==1.2.4",
]